+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auto_classes/auto1.java
@@ -1,12 +1,15 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.auto_classes;
 
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.GyroSensor;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
-@TeleOp(name="test", group="test") //fix this
-public class teleOpMecanum extends teamMethods {
+import org.firstinspires.ftc.teamcode.team_methods.TeamMethods;
+
+
+@TeleOp(name="test2", group="test") //fix this
+public class auto1 extends TeamMethods {
     private ElapsedTime runtime = new ElapsedTime();
 
     //these values should be determined based off hardware being used
@@ -63,23 +66,10 @@ public class teleOpMecanum extends teamMethods {
     //Initialized by: After Start, Before Stop / loops
     @Override
     public void loop() {
-        double drivex = -gamepad1.left_stick_y;
-        double drivey = gamepad1.left_stick_x;
-        double turn  =  gamepad1.right_stick_x;
-        double relativeValues[] = relativeValues(drivex,drivey,turn);
-        double xNew = relativeValues[1];
-        double yNew = relativeValues[2];
-        double angleNew = relativeValues[3];
-        driveToPosition(xNew,yNew,angleNew, true);
-        telemetry.addData("Status", "Run Time: " + runtime.toString());
     }
 
     //Initialized by: Stop / runs once
     @Override
-    public void stop() {
-        motor1.setPower(0);
-        motor2.setPower(0);
-        motor3.setPower(0);
-        motor4.setPower(0);
-    }
+    public void stop() {}
+
 }
